+++ src/shared/util.js
@@ -382,18 +382,6 @@ function shadow(obj, prop, value) {
   return value;
 }
 
-function getLookupTableFactory(initializer) {
-  var lookup;
-  return function () {
-    if (initializer) {
-      lookup = Object.create(null);
-      initializer(lookup);
-      initializer = null;
-    }
-    return lookup;
-  };
-}
-
 var PasswordException = (function PasswordExceptionClosure() {
   function PasswordException(msg, code) {
     this.name = 'PasswordException';
@@ -456,44 +444,6 @@ var UnexpectedResponseException =
   UnexpectedResponseException.constructor = UnexpectedResponseException;
 
   return UnexpectedResponseException;
-})();
-
-var MissingDataException = (function MissingDataExceptionClosure() {
-  function MissingDataException(begin, end) {
-    this.begin = begin;
-    this.end = end;
-    this.message = 'Missing data [' + begin + ', ' + end + ')';
-  }
-
-  MissingDataException.prototype = new Error();
-  MissingDataException.prototype.name = 'MissingDataException';
-  MissingDataException.constructor = MissingDataException;
-
-  return MissingDataException;
-})();
-
-const XRefEntryException = (function XRefEntryExceptionClosure() {
-  function XRefEntryException(msg) {
-    this.message = msg;
-  }
-
-  XRefEntryException.prototype = new Error();
-  XRefEntryException.prototype.name = 'XRefEntryException';
-  XRefEntryException.constructor = XRefEntryException;
-
-  return XRefEntryException;
-})();
-
-var XRefParseException = (function XRefParseExceptionClosure() {
-  function XRefParseException(msg) {
-    this.message = msg;
-  }
-
-  XRefParseException.prototype = new Error();
-  XRefParseException.prototype.name = 'XRefParseException';
-  XRefParseException.constructor = XRefParseException;
-
-  return XRefParseException;
 })();
 
 /**
@@ -659,53 +609,6 @@ function isEvalSupported() {
   }
 }
 
-/**
- * Get the value of an inheritable property.
- *
- * If the PDF specification explicitly lists a property in a dictionary as
- * inheritable, then the value of the property may be present in the dictionary
- * itself or in one or more parents of the dictionary.
- *
- * If the key is not found in the tree, `undefined` is returned. Otherwise,
- * the value for the key is returned or, if `stopWhenFound` is `false`, a list
- * of values is returned. To avoid infinite loops, the traversal is stopped when
- * the loop limit is reached.
- *
- * @param {Dict} dict - Dictionary from where to start the traversal.
- * @param {string} key - The key of the property to find the value for.
- * @param {boolean} getArray - Whether or not the value should be fetched as an
- *   array. The default value is `false`.
- * @param {boolean} stopWhenFound - Whether or not to stop the traversal when
- *   the key is found. If set to `false`, we always walk up the entire parent
- *   chain, for example to be able to find `\Resources` placed on multiple
- *   levels of the tree. The default value is `true`.
- */
-function getInheritableProperty({ dict, key, getArray = false,
-                                  stopWhenFound = true, }) {
-  const LOOP_LIMIT = 100;
-  let loopCount = 0;
-  let values;
-
-  while (dict) {
-    const value = getArray ? dict.getArray(key) : dict.get(key);
-    if (value !== undefined) {
-      if (stopWhenFound) {
-        return value;
-      }
-      if (!values) {
-        values = [];
-      }
-      values.push(value);
-    }
-    if (++loopCount > LOOP_LIMIT) {
-      warn(`getInheritableProperty: maximum loop count exceeded for "${key}"`);
-      break;
-    }
-    dict = dict.get('Parent');
-  }
-  return values;
-}
-
 var Util = (function UtilClosure() {
   function Util() {}
 
@@ -866,43 +769,6 @@ var Util = (function UtilClosure() {
   return Util;
 })();
 
-const ROMAN_NUMBER_MAP = [
-  '', 'C', 'CC', 'CCC', 'CD', 'D', 'DC', 'DCC', 'DCCC', 'CM',
-  '', 'X', 'XX', 'XXX', 'XL', 'L', 'LX', 'LXX', 'LXXX', 'XC',
-  '', 'I', 'II', 'III', 'IV', 'V', 'VI', 'VII', 'VIII', 'IX'
-];
-
-/**
- * Converts positive integers to (upper case) Roman numerals.
- * @param {integer} number - The number that should be converted.
- * @param {boolean} lowerCase - Indicates if the result should be converted
- *   to lower case letters. The default value is `false`.
- * @return {string} The resulting Roman number.
- */
-function toRomanNumerals(number, lowerCase = false) {
-  assert(Number.isInteger(number) && number > 0,
-         'The number should be a positive integer.');
-  let pos, romanBuf = [];
-  // Thousands
-  while (number >= 1000) {
-    number -= 1000;
-    romanBuf.push('M');
-  }
-  // Hundreds
-  pos = (number / 100) | 0;
-  number %= 100;
-  romanBuf.push(ROMAN_NUMBER_MAP[pos]);
-  // Tens
-  pos = (number / 10) | 0;
-  number %= 10;
-  romanBuf.push(ROMAN_NUMBER_MAP[10 + pos]);
-  // Ones
-  romanBuf.push(ROMAN_NUMBER_MAP[20 + number]);
-
-  const romanStr = romanBuf.join('');
-  return (lowerCase ? romanStr.toLowerCase() : romanStr);
-}
-
 const PDFStringTranslateTable = [
   0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
   0x2D8, 0x2C7, 0x2C6, 0x2D9, 0x2DD, 0x2DB, 0x2DA, 0x2DC, 0, 0, 0, 0, 0, 0, 0,
@@ -1046,7 +912,6 @@ export {
   CMapCompressionType,
   AbortException,
   InvalidPDFException,
-  MissingDataException,
   MissingPDFException,
   NativeImageDecoding,
   PasswordException,
@@ -1057,9 +922,6 @@ export {
   UnexpectedResponseException,
   UnknownErrorException,
   Util,
-  toRomanNumerals,
-  XRefEntryException,
-  XRefParseException,
   FormatError,
   arrayByteLength,
   arraysToBytes,
@@ -1068,8 +930,6 @@ export {
   createPromiseCapability,
   createObjectURL,
   deprecated,
-  getInheritableProperty,
-  getLookupTableFactory,
   getVerbosityLevel,
   info,
   isArrayBuffer,
